refactor: isolate Popover stories
